post.mentioned_accounts: operate on rendered text
The renderer marks up everything that looks like a valid JID or username
mention (this automatically excludes mentions in code blocks, for
example), so all that has to be done is validate the JIDs or usernames.
We operate on the rendered representation to avoid regular expression
that would also extract mentions inside of code blocks.